git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActiveScheduler@5809 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ scheduler-plugins-src/org.objectweb.proactive.scheduler.plugin/src/org/objectweb/proactive/extra/scheduler/gui/actions/KillJobAction.java
@@ -1,7 +1,9 @@
 package org.objectweb.proactive.extra.scheduler.gui.actions;
 
 import org.eclipse.jface.action.Action;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.resource.ImageDescriptor;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.TableItem;
 import org.objectweb.proactive.extra.scheduler.gui.data.SchedulerProxy;
 import org.objectweb.proactive.extra.scheduler.gui.data.TableManager;
@@ -12,8 +14,10 @@ public class KillJobAction extends Action {
 	public static final boolean ENABLED_AT_CONSTRUCTION = false;
 
 	private static KillJobAction instance = null;
+	private Shell shell = null;
 
-	private KillJobAction() {
+	private KillJobAction(Shell shell) {
+		this.shell = shell;
 		this.setText("Kill job");
 		this.setToolTipText("To kill a job (this will remove this job from the scheduler)");
 		this.setImageDescriptor(ImageDescriptor.createFromFile(this.getClass(), "icons/kill.png"));
@@ -22,21 +26,22 @@ public class KillJobAction extends Action {
 
 	@Override
 	public void run() {
-		TableItem item = TableManager.getInstance().getLastSelectedItem();
-		if (item != null) {
-			JobId jobId = (JobId) item.getData();
-			SchedulerProxy.getInstance().kill(jobId);
+		if (MessageDialog.openConfirm(shell, "Confirm please",
+				"Are you sure you want to Kill the scheduler ?")) {
+			TableItem item = TableManager.getInstance().getLastSelectedItem();
+			if (item != null) {
+				JobId jobId = (JobId) item.getData();
+				SchedulerProxy.getInstance().kill(jobId);
+			}
 		}
 	}
 
-	public static KillJobAction newInstance() {
-		instance = new KillJobAction();
+	public static KillJobAction newInstance(Shell shell) {
+		instance = new KillJobAction(shell);
 		return instance;
 	}
 
 	public static KillJobAction getInstance() {
-		if (instance == null)
-			instance = new KillJobAction();
 		return instance;
 	}
 }
